+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ### UNRELEASED
 
+ - Add optional support for newrelic on aioredis data source
+
  - Fix aoredis data sources
 
 ### 0.22.9 - 2020-09-08